backend/requirements.txt
# StandupAI Backend Dependencies

# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Pydantic
pydantic==2.5.2
pydantic-settings==2.1.0

# HTTP Client
httpx==0.25.2

# WebSocket Client (for ElevenLabs ConvAI)
websockets==12.0

# Scheduler
APScheduler==3.10.4

# JWT
PyJWT==2.8.0
cryptography==41.0.7

# Email
aiosmtplib==3.0.1

# Environment
python-dotenv==1.0.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
